+++ library-appcompat/src/main/java/com/cz/android/sample/library/adapter/SimpleArrayAdapter.java
@@ -10,6 +10,9 @@ import android.view.ViewGroup;
 import android.widget.TextView;
 
 
+import com.cz.android.sample.library.data.DataManager;
+import com.cz.android.sample.library.data.DataProvider;
+
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -18,10 +21,26 @@ import java.util.List;
  * @author Created by cz
  * @date 2020-01-28 18:37
  * @email bingo110@126.com
+ * A simple array adapter for the RecyclerView.
+ *
+ * See the methods below to create an easy adapter for your list.
+ * @see #createFromDataProvider(Context)
+ * @see #createFromDataProvider(Context, int)
+ * @see #createFromResource(Context, int)
  */
 public class SimpleArrayAdapter<E> extends RecyclerView.Adapter<RecyclerView.ViewHolder> {
     public static SimpleArrayAdapter createFromResource(Context context, @ArrayRes int res){
         return new SimpleArrayAdapter(context, context.getResources().getStringArray(res));
+    }
+
+    public static SimpleArrayAdapter createFromDataProvider(Context context){
+        DataProvider dataProvider = DataManager.getDataProvider(context);
+        return new SimpleArrayAdapter(context, dataProvider.getItemArray());
+    }
+
+    public static SimpleArrayAdapter createFromDataProvider(Context context,int length){
+        DataProvider dataProvider = DataManager.getDataProvider(context);
+        return new SimpleArrayAdapter(context, dataProvider.getItemArray(length));
     }
 
     private final LayoutInflater layoutInflater;