+++ library-appcompat/src/main/java/com/cz/android/sample/library/generate/SampleCategoryGenerator.java
@@ -9,5 +9,5 @@ import com.cz.android.sample.api.item.RegisterItem;
 import java.util.List;
 
 public interface SampleCategoryGenerator {
-    List<CategoryItem> generate(@NonNull Context context, @NonNull List<RegisterItem> registerItemList);
+    List<CategoryItem> generate(@NonNull Context context, @NonNull List<RegisterItem> registeredItemList,@NonNull List<RegisterItem> newItemList);
 }