+++ library-appcompat/src/main/java/com/cz/android/sample/library/main/AndroidSampleSupport.java
@@ -1,9 +1,11 @@
 package com.cz.android.sample.library.main;
 
+import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentActivity;
 
 import com.cz.android.sample.AndroidSample;
 import com.cz.android.sample.function.FunctionManager;
+import com.cz.android.sample.main.MainSampleComponentFactory;
 
 /**
  * @author Created by cz
@@ -14,4 +16,11 @@ public interface AndroidSampleSupport extends AndroidSample<FragmentActivity> {
     String TAG="AndroidSampleSupport";
 
     FunctionManager getFunctionManager();
+
+    /**
+     * Return the component factory.
+     * @see MainSampleComponentFactory
+     * @return
+     */
+    MainSampleComponentFactory<Fragment> getComponentContainerFactory();
 }