+++ library/src/main/java/com/cz/android/sample/library/adapter/SimpleArrayAdapter.java
@@ -10,6 +10,9 @@ import androidx.annotation.LayoutRes;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.cz.android.sample.library.data.DataManager;
+import com.cz.android.sample.library.data.DataProvider;
+
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -18,21 +21,38 @@ import java.util.List;
  * @author Created by cz
  * @date 2020-01-28 18:37
  * @email bingo110@126.com
+ * A simple array adapter for the RecyclerView.
+ *
+ * See the methods below to create an easy adapter for your list.
+ * @see #createFromDataProvider(Context)
+ * @see #createFromDataProvider(Context, int)
+ * @see #createFromResource(Context, int)
  */
 public class SimpleArrayAdapter<E> extends RecyclerView.Adapter<RecyclerView.ViewHolder> {
     public static SimpleArrayAdapter createFromResource(Context context, @ArrayRes int res){
         return new SimpleArrayAdapter(context, context.getResources().getStringArray(res));
     }
 
+    public static SimpleArrayAdapter createFromDataProvider(Context context){
+        DataProvider dataProvider = DataManager.getDataProvider(context);
+        return new SimpleArrayAdapter(context, dataProvider.getItemArray());
+    }
+
+    public static SimpleArrayAdapter createFromDataProvider(Context context,int length){
+        DataProvider dataProvider = DataManager.getDataProvider(context);
+        return new SimpleArrayAdapter(context, dataProvider.getItemArray(length));
+    }
+
     private final LayoutInflater layoutInflater;
-    private @LayoutRes int layoutResources;
+    @LayoutRes
+    private int layoutResources;
     private List<E> items= new ArrayList<>();
 
     public SimpleArrayAdapter(Context context, E[] items){
         this(context, android.R.layout.simple_list_item_1, Arrays.asList(items));
     }
 
-    public SimpleArrayAdapter(Context context, @LayoutRes int layout,E[]  items){
+    public SimpleArrayAdapter(Context context, @LayoutRes int layout, E[]  items){
         this(context, layout, Arrays.asList(items));
     }
 
@@ -40,7 +60,7 @@ public class SimpleArrayAdapter<E> extends RecyclerView.Adapter<RecyclerView.Vie
         this(context,android.R.layout.simple_list_item_1,items);
     }
 
-    public SimpleArrayAdapter(Context context,@LayoutRes int layout,@NonNull List<E> items){
+    public SimpleArrayAdapter(Context context, @LayoutRes int layout, @NonNull List<E> items){
         this.layoutInflater = LayoutInflater.from(context);
         this.layoutResources = layout;
         this.items.addAll(items);