+++ library/src/main/java/com/cz/android/sample/library/component/code/SampleSourceCodeComponent.java
@@ -1,10 +1,12 @@
 package com.cz.android.sample.library.component.code;
 
+import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
 import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentActivity;
 import androidx.fragment.app.FragmentPagerAdapter;
@@ -35,7 +37,7 @@ public class SampleSourceCodeComponent extends CompanionComponentContainer<Fragm
     }
 
     @Override
-    public View onCreateCompanionComponent(@NonNull FragmentActivity context, @NonNull Object object, @NonNull ViewGroup parentView, @NonNull View view) {
+    public View onCreateCompanionComponent(@NonNull FragmentActivity context, @NonNull Object object, @NonNull ViewGroup parentView, @NonNull View view, Bundle saveInstance) {
         LayoutInflater layoutInflater = LayoutInflater.from(context);
         View contentLayout = layoutInflater.inflate(R.layout.sample_fragment_tab, parentView, false);
         ViewPager sampleViewPager=contentLayout.findViewById(R.id.sampleViewPager);
@@ -54,7 +56,7 @@ public class SampleSourceCodeComponent extends CompanionComponentContainer<Fragm
     }
 
     @Override
-    public View getComponentView(@NonNull FragmentActivity context, @NonNull Object object, @NonNull ViewGroup parentView, @NonNull View view) {
+    public View getComponentView(@NonNull FragmentActivity context, @NonNull Object object, @NonNull ViewGroup parentView, @NonNull View view, Bundle saveInstance) {
         ViewPager sampleViewPager=view.findViewById(R.id.sampleViewPager);
         TabLayout sampleTabLayout=view.findViewById(R.id.sampleTabLayout);
         PagerAdapter adapter = sampleViewPager.getAdapter();