+++ library/src/main/java/com/cz/android/sample/library/component/message/SampleMessageComponent.java
@@ -1,5 +1,6 @@
 package com.cz.android.sample.library.component.message;
 
+import android.os.Bundle;
 import android.text.Editable;
 import android.text.TextWatcher;
 import android.view.LayoutInflater;
@@ -41,7 +42,7 @@ public class SampleMessageComponent implements ComponentContainer <FragmentActiv
     }
 
     @Override
-    public View getComponentView(FragmentActivity context, Object object, ViewGroup parentView,View view) {
+    public View getComponentView(FragmentActivity context, Object object, ViewGroup parentView,View view, Bundle saveInstance) {
         LayoutInflater layoutInflater = LayoutInflater.from(context);
         View contentLayout = layoutInflater.inflate(R.layout.sample_message_layout, parentView, false);
         FrameLayout sampleMessageContentLayout=contentLayout.findViewById(R.id.sampleMessageContentLayout);