+++ library/src/main/java/com/cz/android/sample/library/generate/SampleItemGenerator.java
@@ -14,5 +14,6 @@ import java.util.List;
  * @email bingo110@126.com
  */
 public interface SampleItemGenerator {
+    @NonNull
     List<RegisterItem> generate(@NonNull Context context, @NonNull List<String> sampleClassList);
 }